refactor(ts): convert pages to TypeScript
Type _app, _document, and the sign-in page's server props. Narrow
router query values before passing them as error messages, and give
the non-PR contributions API response a precise record type so the
dashboard can read RepoName, RepoUrl, and created_at without casts.
The dashboard now guards on the query's data rather than its status,
which also avoids dereferencing a null response.
