+++ pages/_app.tsx
@@ -2,13 +2,14 @@ import { SessionProvider } from 'next-auth/react'
 import '../styles/globals.css'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import Head from 'next/head'
+import type { AppProps } from 'next/app'
 
 // Create a client
 const queryClient = new QueryClient()
 
 // Use the <Provider> to improve performance and allow components that call
 // `useSession()` anywhere in your application to access the `session` object.
-export default function App({ Component, pageProps }) {
+export default function App({ Component, pageProps }: AppProps) {
 	return (
 		<>
 			<Head>
