+++ pages/_document.tsx
@@ -1,7 +1,13 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document'
+import Document, {
+	type DocumentContext,
+	Html,
+	Head,
+	Main,
+	NextScript,
+} from 'next/document'
 
 class MyDocument extends Document {
-	static async getInitialProps(ctx) {
+	static async getInitialProps(ctx: DocumentContext) {
 		const initialProps = await Document.getInitialProps(ctx)
 		return { ...initialProps }
 	}
