+++ pages/auth/signin.tsx
@@ -1,10 +1,13 @@
 import { getProviders, signIn } from 'next-auth/react'
+import type { InferGetServerSidePropsType } from 'next'
 import Layout from '../../components/Layout'
 
-export default function SignIn({ providers }) {
+export default function SignIn({
+	providers,
+}: InferGetServerSidePropsType<typeof getServerSideProps>) {
 	return (
 		<Layout>
-			{Object.values(providers).map((provider) => (
+			{Object.values(providers ?? {}).map((provider) => (
 				<div key={provider.name}>
 					<button onClick={() => signIn(provider.id)}>
 						Sign in with {provider.name}
@@ -16,7 +19,7 @@ export default function SignIn({ providers }) {
 }
 
 // This is the recommended way for Next.js 9.3 or newer
-export async function getServerSideProps(context) {
+export async function getServerSideProps() {
 	const providers = await getProviders()
 	return {
 		props: { providers },
