+++ pages/maintainers.tsx
@@ -1,10 +1,13 @@
 import { useRouter } from 'next/router'
 import { useSession } from 'next-auth/react'
+import Head from 'next/head'
+import Layout from '../components/Layout'
 import Form, { FormLayout } from '../components/Forms'
 import SignIn from '../components/SignIn'
 import { useQuery } from '@tanstack/react-query'
 import Button from '../components/Button'
-import { getMentorsSubmission } from '../util/api'
+import { getMaintainersSubmission } from '../util/api'
+import { currentYear, useNewSubmissionsClosed } from '../util/globals'
 
 // Become a Contributor: Virtual Coffee Hacktoberfest Initiative
 
@@ -12,13 +15,13 @@ const intro = (
 	<>
 		<div className="text-center">
 			<h1 className="text-3xl leading-9 font-extrabold tracking-tight text-gray-900 sm:text-4xl sm:leading-10">
-				Become a Hacktoberfest Mentor
+				Calling All Maintainers
 			</h1>
 			<p className="mt-4 text-lg leading-6 text-gray-500">
-				Thank you for expressing your interest in mentoring a Virtual Coffee
-				member participating the Virtual Coffee Hacktoberfest Initiative. We
-				would really appreciate it if you took the time to fill out this short
-				questionnaire.
+				Are you an open source maintainer who is interested in participating in
+				Hacktoberfest? We're looking for some OSS maintainers to partner with in
+				order to provide a welcoming environment to our contributors as they
+				start their open source journey.
 			</p>
 		</div>
 	</>
@@ -52,10 +55,11 @@ export default function Page() {
 	const { data: session, status: sessionStatus } = useSession()
 	const router = useRouter()
 	const { error, message: errorMessage } = router.query
+	const newSubmissionsClosed = useNewSubmissionsClosed()
 
 	const previousFormSubmission = useQuery({
-		queryKey: ['mentors-form'],
-		queryFn: getMentorsSubmission,
+		queryKey: ['maintainers-form'],
+		queryFn: getMaintainersSubmission,
 		enabled: sessionStatus === 'authenticated',
 	})
 
@@ -66,8 +70,8 @@ export default function Page() {
 	if (sessionStatus === 'unauthenticated') {
 		return (
 			<FormLayout
-				title="Become a Hacktoberfest Mentor"
-				description="Virtual Coffee's Hacktoberfest Initiative is a great place to provide high-impact help to a few early-career contributors."
+				title="Calling All Maintainers"
+				description="We're looking for some OSS maintainers to partner with in order to provide a welcoming environment to our contributors as they start their open source journey."
 			>
 				<SignIn />
 			</FormLayout>
@@ -78,10 +82,32 @@ export default function Page() {
 		return null
 	}
 
+	if (!previousFormSubmission.data && newSubmissionsClosed) {
+		return (
+			<FormLayout
+				title={`We've ended submissions for ${currentYear}`}
+				description="Thank you for your interest, but we've ended submissions for this year. See you next year!"
+			>
+				<div className="text-center">
+					<h1 className="text-3xl leading-9 font-extrabold tracking-tight text-gray-900 sm:text-4xl sm:leading-10">
+						We've ended submissions for {currentYear}
+					</h1>
+					<p className="mt-4 text-lg leading-6 text-gray-500">
+						Thank you for your interest, but we've ended submissions for this
+						year.
+					</p>
+					<p className="mt-4 text-lg leading-6 text-gray-500">
+						See you next year!
+					</p>
+				</div>
+			</FormLayout>
+		)
+	}
+
 	return (
 		<FormLayout
-			title="Become a Hacktoberfest Mentor"
-			description="Virtual Coffee's Hacktoberfest Initiative is a great place to provide high-impact help to a few early-career contributors."
+			title="Calling All Maintainers"
+			description="We're looking for some OSS maintainers to partner with in order to provide a welcoming environment to our contributors as they start their open source journey."
 		>
 			<Form
 				session={session}
@@ -90,11 +116,13 @@ export default function Page() {
 						? previousFormSubmission.data.fields
 						: null
 				}
-				errorMessage={error ? errorMessage : undefined}
+				errorMessage={
+					error && typeof errorMessage === 'string' ? errorMessage : undefined
+				}
 				successView={successView}
 				intro={intro}
-				formKey="mentors"
-				fieldsetLegend="Mentor Details"
+				formKey="maintainers"
+				fieldsetLegend="Maintainer Details"
 			/>
 		</FormLayout>
 	)
