+++ pages/non-pr-contributions.tsx
@@ -1,12 +1,10 @@
 import { useRouter } from 'next/router'
 import { useSession } from 'next-auth/react'
-import Head from 'next/head'
-import Layout from '../components/Layout'
 import Form, { FormLayout } from '../components/Forms'
 import SignIn from '../components/SignIn'
 import { useQuery } from '@tanstack/react-query'
 import Button from '../components/Button'
-import { getMaintainersSubmission } from '../util/api'
+import { getContributorSubmission } from '../util/api'
 import { currentYear, useNewSubmissionsClosed } from '../util/globals'
 
 // Become a Contributor: Virtual Coffee Hacktoberfest Initiative
@@ -15,31 +13,34 @@ const intro = (
 	<>
 		<div className="text-center">
 			<h1 className="text-3xl leading-9 font-extrabold tracking-tight text-gray-900 sm:text-4xl sm:leading-10">
-				Calling All Maintainers
+				Add a Non-PR Contribution
 			</h1>
 			<p className="mt-4 text-lg leading-6 text-gray-500">
-				Are you an open source maintainer who is interested in participating in
-				Hacktoberfest? We're looking for some OSS maintainers to partner with in
-				order to provide a welcoming environment to our contributors as they
-				start their open source journey.
+				Not all open source contributions need to be code! If you have
+				contributed to open source this month in other ways, please tell us
+				about it!
 			</p>
 		</div>
 	</>
 )
 
+{
+	/* <Head>
+			<title>Thank you for your interest!</title>
+		</Head> */
+}
 const successView = (
 	<div className="py-16 px-4  sm:px-6 lg:px-8 lg:py-24">
 		<div className="relative max-w-2xl mx-auto">
 			<div className="text-center">
 				<h1 className="text-3xl leading-9 font-extrabold tracking-tight text-gray-900 sm:text-4xl sm:leading-10 flex justify-center">
 					<span className="mr-4">🙌</span>
-					<span>Thank you for your interest!</span>
+					<span>Thank you for your contribution!</span>
 					<span className="ml-4">🙌</span>
 				</h1>
 				<div className="text-lg leading-6 text-gray-500">
-					<p className="mt-4">
-						You'll hear from a Virtual Coffee member soon with next steps.
-					</p>
+					<p className="mt-4">You're a star!</p>
+
 					<p className="mt-4">
 						<Button size="lg" href="/dashboard">
 							Go to Dashboard
@@ -57,12 +58,6 @@ export default function Page() {
 	const { error, message: errorMessage } = router.query
 	const newSubmissionsClosed = useNewSubmissionsClosed()
 
-	const previousFormSubmission = useQuery({
-		queryKey: ['maintainers-form'],
-		queryFn: getMaintainersSubmission,
-		enabled: sessionStatus === 'authenticated',
-	})
-
 	if (sessionStatus === 'loading') {
 		return null
 	}
@@ -70,19 +65,15 @@ export default function Page() {
 	if (sessionStatus === 'unauthenticated') {
 		return (
 			<FormLayout
-				title="Calling All Maintainers"
-				description="We're looking for some OSS maintainers to partner with in order to provide a welcoming environment to our contributors as they start their open source journey."
+				title="Add a Non-PR Contribution"
+				description="Not all open source contributions need to be code!"
 			>
 				<SignIn />
 			</FormLayout>
 		)
 	}
 
-	if (!(previousFormSubmission.isSuccess || previousFormSubmission.isError)) {
-		return null
-	}
-
-	if (!previousFormSubmission.data && newSubmissionsClosed) {
+	if (newSubmissionsClosed) {
 		return (
 			<FormLayout
 				title={`We've ended submissions for ${currentYear}`}
@@ -106,21 +97,20 @@ export default function Page() {
 
 	return (
 		<FormLayout
-			title="Calling All Maintainers"
-			description="We're looking for some OSS maintainers to partner with in order to provide a welcoming environment to our contributors as they start their open source journey."
+			title="Add a Non-PR Contribution"
+			description="Not all open source contributions need to be code!"
 		>
 			<Form
+				showProfileFields={false}
 				session={session}
-				previousFormSubmission={
-					previousFormSubmission?.data?.success
-						? previousFormSubmission.data.fields
-						: null
+				errorMessage={
+					error && typeof errorMessage === 'string' ? errorMessage : undefined
 				}
-				errorMessage={error ? errorMessage : undefined}
 				successView={successView}
 				intro={intro}
-				formKey="maintainers"
-				fieldsetLegend="Maintainer Details"
+				formKey="nonPrContributions"
+				fieldsetLegend="Contribution Details"
+				submitText="Save Contribution"
 			/>
 		</FormLayout>
 	)
